requirements.txt
asgiref==3.5.0
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
coverage==6.3.1
Django==4.0.1
django-extensions==3.1.5
django-filter==21.1
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==12.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
packaging==21.3
pep8==1.7.1
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==7.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
